feat(src): add HeroSection component to Home page
